Refactor oscillating droplet benchmark operators
