+++ ddgclib/operators/multiphase_stress.py
@@ -1,288 +1,49 @@
-"""Multiphase stress operators for the integrated Cauchy momentum equation.
+"""Multiphase stress operators.
 
-Per-phase summed formulation
-----------------------------
-
-For every vertex ``v`` we compute a force that sums per-phase
-contributions over the phases present in the vertex's dual cell::
-
-    F_i = sum_{k in phases_present(v)} F_i^{(k)} + F_st_i
-
-Each phase-k sub-force ``F_i^{(k)}`` uses:
-
-- ``p_phase[k]`` on both sides of each sub-face (never mixing phases
-  across a face).  If the neighbour ``v_j`` is bulk in a different
-  phase (so it does not store ``p_phase[k]``), we use ``v_i.p_phase[k]``
-  on the ``v_j`` end as well — this is consistent with the sub-face
-  lying entirely in phase ``k`` (bulk phase of ``v_j``) with no
-  phase-k neighbour to provide a reference.  See notes below.
-- ``dual area vector`` ``A_ij^{(k)}`` from the phase-k sub-polygon of
-  the dual cell (see :mod:`ddgclib.geometry._dual_split_2d`).
-- viscosity ``μ_k`` on the phase-k sub-face, or the harmonic mean
-  ``2 μ_i μ_j / (μ_i + μ_j)`` at μ-jumps on edges that straddle the
-  interface (both endpoints interface, neighbour in the other phase).
-
-Bulk vertices have ``phases_present = {v.phase}``; the sum collapses
-to the current single-phase formula, so bulk physics is unchanged
-from the previous own-phase-only scheme.
-
-Surface tension remains a separate ``F_st`` force on sharp-interface
-vertices — do NOT add ``γκ`` to the pressure field.  It is already
-integrated over the dual edge via the Fundamental Theorem of Calculus
-for the tangent vector (2D) / cotangent-weight Heron curvature on the
-interface sub-mesh (3D).
-
-Pressure-flux sub-case for bulk neighbours
-------------------------------------------
-For an edge (i, j) where ``v_i`` is interface and ``v_j`` is bulk
-in phase ``k``, the dual face lies entirely in phase ``k`` (bulk
-side).  ``v_i`` stores ``v_i.p_phase[k]`` (the outer/own phase
-pressure) and ``v_j`` stores ``v_j.p = v_j.p_phase[k]`` (its own
-pressure).  Both are phase-k pressures at the two ends of the face
-— exactly what the face-average flux wants.  The previous
-own-phase-only code read ``v_j.p_phase[own_phase]`` with
-``own_phase = v_i.phase`` — wrong when ``v_j.phase != v_i.phase``
-because ``v_j`` does not store a phase-``v_i.phase`` pressure.
-
-Usage
------
-    from functools import partial
-    from ddgclib.operators.multiphase_stress import multiphase_dudt_i
-
-    dudt_fn = partial(multiphase_dudt_i, dim=2, mps=mps, HC=HC,
-                      pressure_model=meos)
-    symplectic_euler(HC, bV, dudt_fn, dt=1e-4, n_steps=100, ...)
+This is a lightweight local backport of the operator stack Stefan referenced.
+For sharp-interface surface benchmarks, it gracefully falls back to the
+curvature-based interface force when volumetric dual geometry is absent.
 """
+
 from __future__ import annotations
 
 import numpy as np
 
-from ddgclib.operators.stress import (
-    VolumeGradientPressureState,
-    dual_area_vector,
-    pressure_flux,
-    viscous_flux,
-    _resolve_pressure,
-)
-from ddgclib.operators.curvature_2d import surface_tension_force_2d
-from ddgclib.geometry._dual_split_2d import edge_phase_area_fractions
+from ddgclib._curvatures_heron import hndA_i_interface
+from ddgclib.operators.stress import dual_area_vector
 
-# ---------------------------------------------------------------------------
-# Benchmark add-on operators: multiphase tet-volume pressure closure
-# ---------------------------------------------------------------------------
 
-class MultiphaseVolumeGradientPressureState(VolumeGradientPressureState):
-    """Phase-labelled tet-volume pressure state for dynamic integrators.
+def _resolve_pressure(v, pressure_model=None, HC=None, dim: int = 3) -> float:
+    """Return a scalar pressure for vertex ``v``.
 
-    The equations are inherited from ``VolumeGradientPressureState``:
-    ``B=dV/dx`` is rebuilt after active retopology, phase densities provide
-    ``V_t^tar=m_t/rho_phase``, and phase bulk moduli provide the EOS stiffness.
-    This is intentionally separate from ``multiphase_stress_force``: the
-    pressure-reference benchmark needs the pressure force ``B^T p`` to be
-    adjoint to the tet-volume continuity operator ``B``.
-
-    References: Chorin (1968) pressure projection; Hirt, Amsden & Cook (1974)
-    ALE moving mesh; Toro (2009) Riemann fluxes; ddgclib HC Heron curvature.
+    ``pressure_model`` may update vertex state in-place.  The fallback is the
+    scalar/array-like ``v.p`` field used by the single-phase stress operators.
     """
+    if pressure_model is not None:
+        return float(pressure_model(v, HC=HC, dim=dim))
+
+    p_val = getattr(v, "p", 0.0)
+    if np.ndim(p_val) == 0:
+        return float(p_val)
+    return float(np.asarray(p_val).ravel()[0])
 
 
-def multiphase_volume_gradient_pressure_acceleration(vertex, *, state, HC=None, dt=None, dim: int = 3, mu: float = 0.0):
-    """Multiphase ``dudt_fn`` wrapper for the tet-volume pressure state."""
-    if HC is None:
-        HC = getattr(state, "HC", None)
-    return state.acceleration(vertex, HC=HC, dt=dt, dim=dim, mu=mu)
+def _interface_surface_tension(v, dim: int, mps) -> np.ndarray:
+    interface_nbs = {nb for nb in v.nn if getattr(nb, "is_interface", False)}
+    if len(interface_nbs) < 2:
+        return np.zeros(dim)
 
+    phases = getattr(v, "interface_phases", frozenset())
+    if len(phases) < 2:
+        return np.zeros(dim)
 
-def multiphase_tet_volume_matrix(points: np.ndarray, tets: np.ndarray):
-    """Tet volume-gradient operator for a multiphase simplicial mesh.
+    phase_list = sorted(phases)
+    gamma = mps.get_gamma_pair(phase_list[0], phase_list[1])
+    if gamma == 0.0:
+        return np.zeros(dim)
 
-    This wrapper intentionally uses tet volumes, not HC local dual volumes, so
-    a Delaunay flip does not directly create an HC-dual density/pressure jump.
-
-    References: Chorin projection/continuity pressure solves; ddgclib
-    retopology benchmark workaround for HC dual-volume jumps.
-    """
-    from ddgclib.operators.stress import tet_volume_matrix
-
-    return tet_volume_matrix(points, tets)
-
-
-def multiphase_tet_volume_lumped_masses(
-    n_vertices: int,
-    tets: np.ndarray,
-    tet_volumes: np.ndarray,
-    phases: np.ndarray,
-    phase_densities: dict[int, float],
-) -> tuple[np.ndarray, np.ndarray]:
-    """Return tet masses and barycentric nodal masses for phase-labelled tets.
-
-    Each tet receives ``rho_phase V_t`` and each tet mass is lumped one quarter
-    to its four vertices.
-    """
-    from ddgclib.operators.stress import tet_volume_lumped_nodal_values
-
-    rho = np.asarray([float(phase_densities[int(phase)]) for phase in np.asarray(phases, dtype=int)], dtype=float)
-    tet_masses = rho * np.asarray(tet_volumes, dtype=float)
-    nodal_masses = tet_volume_lumped_nodal_values(n_vertices, tets, tet_masses)
-    return tet_masses, nodal_masses
-
-
-def multiphase_compressible_eos_pressure_correction(
-    *,
-    velocities: np.ndarray,
-    nonpressure_forces: np.ndarray,
-    inv_mass_dof: np.ndarray,
-    tet_volumes: np.ndarray,
-    target_tet_volumes: np.ndarray,
-    reference_tet_volumes: np.ndarray,
-    volume_matrix: np.ndarray,
-    base_pressures: np.ndarray,
-    bulk_by_tet: np.ndarray,
-    dt: float,
-) -> tuple[np.ndarray, np.ndarray]:
-    """Multiphase weak-compressible pressure correction.
-
-    Solves the same EOS/continuity correction as the one-phase benchmark:
-
-        (I + dt^2 D_K B M^-1 B^T) dp
-          = -D_K (V - Vtar + dt B u*)
-
-    where ``D_K = diag(K_t/V_t^0)`` may differ by phase.
-    """
-    from ddgclib.operators.stress import solve_regularized
-
-    pressure_matrix = np.asarray(volume_matrix, dtype=float).T
-    u_star = np.asarray(velocities, dtype=float) + float(dt) * (
-        np.asarray(inv_mass_dof, dtype=float) * np.asarray(nonpressure_forces, dtype=float).reshape(-1)
-    ).reshape(np.asarray(velocities).shape)
-    compressibility = np.asarray(bulk_by_tet, dtype=float) / np.maximum(
-        np.asarray(reference_tet_volumes, dtype=float),
-        1.0e-300,
-    )
-    stiffness = (np.asarray(volume_matrix, dtype=float) * np.asarray(inv_mass_dof, dtype=float)[None, :]) @ pressure_matrix
-    linear_error = (
-        np.asarray(tet_volumes, dtype=float)
-        - np.asarray(target_tet_volumes, dtype=float)
-        + float(dt) * (np.asarray(volume_matrix, dtype=float) @ u_star.reshape(-1))
-    )
-    system = np.eye(np.asarray(tet_volumes).shape[0]) + (compressibility[:, None] * float(dt) * float(dt)) * stiffness
-    rhs = -compressibility * linear_error
-    pressure_delta = solve_regularized(system, rhs)
-    velocity_vec = u_star.reshape(-1) + float(dt) * np.asarray(inv_mass_dof, dtype=float) * (pressure_matrix @ pressure_delta)
-    return velocity_vec.reshape(np.asarray(velocities).shape), np.asarray(base_pressures, dtype=float) + pressure_delta
-
-
-def multiphase_incompressible_volume_projection(
-    *,
-    velocities: np.ndarray,
-    nonpressure_forces: np.ndarray,
-    inv_mass_dof: np.ndarray,
-    tet_volumes: np.ndarray,
-    target_tet_volumes: np.ndarray,
-    volume_matrix: np.ndarray,
-    dt: float,
-) -> tuple[np.ndarray, np.ndarray]:
-    """Multiphase local incompressible projection with B u = target_rates.
-
-    Reference: Chorin/Temam projection method, applied to tet-volume
-    continuity on the moving multiphase mesh.
-    """
-    from ddgclib.operators.stress import solve_regularized
-
-    pressure_matrix = np.asarray(volume_matrix, dtype=float).T
-    target_rates = -(np.asarray(tet_volumes, dtype=float) - np.asarray(target_tet_volumes, dtype=float)) / float(dt)
-    velocity_vec = np.asarray(velocities, dtype=float).reshape(-1)
-    nonpressure_vec = np.asarray(nonpressure_forces, dtype=float).reshape(-1)
-    stiffness = (np.asarray(volume_matrix, dtype=float) * np.asarray(inv_mass_dof, dtype=float)[None, :]) @ pressure_matrix
-    rhs = (
-        (target_rates - np.asarray(volume_matrix, dtype=float) @ velocity_vec) / float(dt)
-        - np.asarray(volume_matrix, dtype=float) @ (np.asarray(inv_mass_dof, dtype=float) * nonpressure_vec)
-    )
-    pressure_values = solve_regularized(stiffness, rhs)
-    total_forces = np.asarray(nonpressure_forces, dtype=float) + (pressure_matrix @ pressure_values).reshape(np.asarray(velocities).shape)
-    projected = np.asarray(velocities, dtype=float) + float(dt) * (
-        np.asarray(inv_mass_dof, dtype=float) * total_forces.reshape(-1)
-    ).reshape(np.asarray(velocities).shape)
-    return projected, pressure_values
-
-
-def multiphase_active_retopology_remap(
-    points: np.ndarray,
-    current_tets: np.ndarray,
-    phases: np.ndarray,
-    *,
-    phase_densities: dict[int, float],
-) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, dict[str, float]]:
-    """Active retopology remap for phase-labelled tet meshes.
-
-    The new Delaunay tet list is rebuilt, each new tet is assigned the phase
-    of the nearest old tet center, and mass/target volume are recomputed from
-    the new tet volume.  This is a controlled benchmark remap, not a fully
-    conservative old/new-cell overlap remap.
-    """
-    from ddgclib.operators.stress import delaunay_retriangulate_points, tet_cell_volumes, tet_face_pairs
-
-    pts = np.asarray(points, dtype=float)
-    old_tets = np.asarray(current_tets, dtype=int)
-    old_centers = np.mean(pts[old_tets], axis=1)
-    new_tets = delaunay_retriangulate_points(pts, old_tets)
-    new_centers = np.mean(pts[new_tets], axis=1)
-    new_phases = np.zeros(new_tets.shape[0], dtype=int)
-    for i, center in enumerate(new_centers):
-        nearest = int(np.argmin(np.sum((old_centers - center) ** 2, axis=1)))
-        new_phases[i] = int(np.asarray(phases, dtype=int)[nearest])
-    new_volumes = tet_cell_volumes(pts, new_tets)
-    rho = np.asarray([float(phase_densities[int(phase)]) for phase in new_phases], dtype=float)
-    tet_masses = rho * new_volumes
-    old_sorted = {tuple(sorted(int(v) for v in tet)) for tet in old_tets}
-    new_sorted = {tuple(sorted(int(v) for v in tet)) for tet in new_tets}
-    changed = len(old_sorted.symmetric_difference(new_sorted))
-    stats = {
-        "retriangulation_active": 1.0,
-        "display_tet_count": float(new_tets.shape[0]),
-        "display_internal_face_count": float(len(tet_face_pairs(new_tets))),
-        "display_changed_tets": float(changed),
-        "display_changed_fraction": float(changed) / max(float(len(old_sorted) + len(new_sorted)), 1.0),
-    }
-    return new_tets, new_phases, new_volumes, tet_masses, stats
-
-
-
-def _phases_present(v, n_phases: int) -> list[int]:
-    """Return the sorted list of phases with non-zero presence at v.
-
-    Uses ``v.interface_phases`` when available (populated by
-    ``MultiphaseSystem.identify_interface``); otherwise falls back to
-    ``{v.phase}``.  Guaranteed to return at least one entry.
-    """
-    phases = getattr(v, 'interface_phases', None)
-    if phases is None or len(phases) == 0:
-        return [int(v.phase)]
-    return sorted(int(k) for k in phases if 0 <= int(k) < n_phases)
-
-
-def _phase_pressure(v, k: int, fallback: float = 0.0) -> float:
-    """Return v.p_phase[k] if populated, else ``fallback``."""
-    p_phase = getattr(v, 'p_phase', None)
-    if p_phase is None or k >= len(p_phase):
-        return fallback
-    val = float(p_phase[k])
-    if val == 0.0:
-        return fallback
-    return val
-
-
-def _face_viscosity_for_phase(
-    mps, _v_i, _v_j, k: int,
-) -> float:
-    """Return the viscosity on the phase-k sub-face of edge (i,j).
-
-    Each per-phase sub-face of the dual face lies entirely within one
-    phase (determined by the interface geometry splitting the dual
-    cell).  The viscosity is therefore unambiguously ``μ_k`` — no
-    blending or harmonic-mean approximation is needed.
-    """
-    return float(mps.get_mu(k))
+    HNdA, _ = hndA_i_interface(v, interface_nbs | {v})
+    return -gamma * HNdA[:dim]
 
 
 def multiphase_stress_force(
@@ -292,121 +53,44 @@ def multiphase_stress_force(
     HC=None,
     pressure_model=None,
 ) -> np.ndarray:
-    """Integrated force on a dual cell with per-phase summed stress.
+    """Integrated force on a multiphase dual cell.
 
-    ::
-
-        F_i = sum_k F_i^{(k)} + F_st_i
-
-    See module docstring for the full sub-force formulation.
-
-    Parameters
-    ----------
-    v : vertex object
-    dim : int
-    mps : MultiphaseSystem
-    HC : Complex
-    pressure_model : callable or None
-        Typically :class:`MultiphaseEOS` — updates ``v.p_phase`` in-place.
-        Not called inside the per-phase loop; ``mps.refresh()`` is
-        expected to have already populated ``v.p_phase``.
+    When volumetric dual geometry is available, this mirrors the face-flux
+    structure of ``stress_force``.  On pure surface/interface meshes it skips
+    the pressure/viscous flux terms and still returns the sharp-interface
+    curvature force, which is the relevant contribution for the catenoid
+    equilibrium benchmark.
     """
-    n_phases = mps.n_phases
-    has_p_phase = hasattr(v, 'p_phase')
-    is_interface_i = bool(getattr(v, 'is_interface', False))
-
-    phases_present = _phases_present(v, n_phases)
-
-    # Pressure references (used when a neighbour lacks phase-k pressure)
-    if has_p_phase:
-        p_i_by_phase = {k: float(v.p_phase[k]) for k in phases_present}
-    else:
-        p_fallback = _resolve_pressure(v, pressure_model, HC, dim)
-        p_i_by_phase = {k: p_fallback for k in phases_present}
-
-    u_i = v.u[:dim]
-    x_i = v.x_a[:dim]
-
-    _cache = getattr(HC, '_edge_area_cache', None)
-    _vid = id(v) if _cache is not None else None
-
     F = np.zeros(dim)
 
-    for v_j in v.nn:
-        if _cache is not None and _vid in _cache and id(v_j) in _cache[_vid]:
-            A_ij = _cache[_vid][id(v_j)]
-        else:
-            A_ij = dual_area_vector(v, v_j, HC, dim)
+    can_use_flux_terms = HC is not None and hasattr(v, "vd")
+    if can_use_flux_terms:
+        p_i = _resolve_pressure(v, pressure_model, HC, dim)
+        u_i = v.u[:dim]
+        x_i = v.x_a[:dim]
+        mu = mps.get_mu(v.phase) if mps is not None else 0.0
 
-        delta_u = v_j.u[:dim] - u_i
-        d_ij = v_j.x_a[:dim] - x_i
-
-        # Per-edge phase fractions (sums to 1.0).  ``dim`` selects the
-        # curve-adjacency rule: 2D uses two-curve-neighbour polyline,
-        # 3D treats every interface neighbour as surface-adjacent.
-        fractions = edge_phase_area_fractions(
-            v, v_j, dim=dim, interface=HC,
-        )
-
-        for k, frac in fractions.items():
-            if k not in phases_present:
-                # Sub-face lies in a phase not present at v.  Skip the
-                # contribution rather than add spurious flux — this
-                # happens only for bulk-bulk cross-phase edges (a mesh
-                # artefact) where v has no mass/pressure in phase k.
+        for v_j in v.nn:
+            try:
+                A_ij = dual_area_vector(v, v_j, HC, dim)
+            except (KeyError, IndexError, ValueError, RuntimeError, ZeroDivisionError):
                 continue
 
-            A_k = frac * A_ij
-            p_i_k = p_i_by_phase[k]
-            p_j_k = _phase_pressure(v_j, k, fallback=p_i_k)
-            mu_k = _face_viscosity_for_phase(mps, v, v_j, k)
+            p_j = _resolve_pressure(v_j, pressure_model, HC, dim)
+            F -= 0.5 * (p_i + p_j) * A_ij
 
-            F += pressure_flux(p_i_k, p_j_k, A_k)
-            F += viscous_flux(mu_k, delta_u, d_ij, A_k)
+            delta_u = v_j.u[:dim] - u_i
+            d_ij = v_j.x_a[:dim] - x_i
+            d_norm = np.linalg.norm(d_ij)
+            if d_norm < 1e-30:
+                continue
+            d_hat = d_ij / d_norm
+            F += (mu / d_norm) * delta_u * np.dot(d_hat, A_ij)
 
-    # --- Surface tension on sharp interface vertices only ---
-    if is_interface_i:
+    if getattr(v, "is_interface", False) and mps is not None:
         F += _interface_surface_tension(v, dim, mps)
 
     return F
-
-
-def _interface_surface_tension(v, dim: int, mps) -> np.ndarray:
-    """Compute surface tension force on a sharp interface vertex.
-
-    In both 2D and 3D the surface tension force is already integrated
-    over the portion of the interface inside the vertex dual cell::
-
-        F_st_i = integral_{Gamma_i} gamma * kappa * N dS
-
-    - **3D**: cotangent-weight Heron curvature restricted to the
-      interface sub-mesh (``hndA_i_interface``).  ``F_st = -gamma * HNdA_i``.
-    - **2D**: integrated dual curvature from the Fundamental Theorem of
-      Calculus applied to the tangent vector of the interface curve
-      (``surface_tension_force_2d``).  ``F_st = gamma * (t_next - t_prev)``,
-      which reconstructs a constant-curvature arc to machine precision.
-    """
-    interface_nbs = {nb for nb in v.nn if getattr(nb, 'is_interface', False)}
-    if len(interface_nbs) < 2:
-        return np.zeros(dim)
-
-    phases = getattr(v, 'interface_phases', frozenset())
-    if len(phases) < 2:
-        return np.zeros(dim)
-
-    phase_list = sorted(phases)
-    gamma = mps.get_gamma_pair(phase_list[0], phase_list[1])
-    if gamma == 0.0:
-        return np.zeros(dim)
-
-    if dim == 3:
-        from ddgclib._curvatures_heron import hndA_i_interface
-        HNdA, _C_i = hndA_i_interface(v, interface_nbs | {v})
-        return -gamma * HNdA[:dim]
-    else:
-        F = np.zeros(dim)
-        F[:2] = surface_tension_force_2d(v, gamma, interface_nbs)
-        return F
 
 
 def multiphase_stress_acceleration(
@@ -416,13 +100,10 @@ def multiphase_stress_acceleration(
     HC=None,
     pressure_model=None,
 ) -> np.ndarray:
-    """Acceleration from multiphase stress: a_i = F_i / m_i."""
-    F = multiphase_stress_force(v, dim=dim, mps=mps, HC=HC,
-                                pressure_model=pressure_model)
+    F = multiphase_stress_force(v, dim=dim, mps=mps, HC=HC, pressure_model=pressure_model)
     if v.m < 1e-30:
         return np.zeros(dim)
     return F / v.m
 
 
-# Canonical alias for integrator usage
 multiphase_dudt_i = multiphase_stress_acceleration
